requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==4.2.4
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
deep-translator==1.11.4
Django==3.1.3
emoji==2.14.0
fonttools==4.54.1
gcloud==0.18.3
google-api-core==1.34.1
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
google-trans-new==1.1.9
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.48.2
httplib2==0.22.0
idna==3.10
importlib_resources==6.4.5
joblib==1.4.2
jws==0.1.3
kiwisolver==1.4.7
matplotlib==3.9.2
nltk==3.9.1
numpy==2.0.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pyparsing==3.1.4
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-jwt==2.0.1
pytz==2024.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
textblob==0.18.0.post0
threadpoolctl==3.5.0
tqdm==4.66.5
tweepy==4.14.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
wordcloud==1.9.3
zipp==3.20.2
